Update performance-best-practices.md
Convert Eligible Microflows to Nanoflows.
Added comment on 2nd bullet. Create actions should NEVER be converted to Nanoflows due to security reasons. For performance reasons converting to a nanoflow is the most BAD practice there is. 
in case of a single object, performance isn't at stake. And security prevails, thus keep microflow.
Incase of many objects creating them on the app server instead of client performs better. Hence the app server is stronger than the average client on this.

TLDR: please rewrite the sentense in such that it excludes any create action. AND ensure this is removed from the Best practice recommender in Studio Pro.